Estimate the SFGD (q0) m2/d.

Estimate the SGD volume based on Strack (1976) analytical solution.

Usage

estimate_sgd(
  inputData,
  calibratableParams,
  constParams,
  nw = 120L,
  warmup = 1500L
)

Arguments

inputData

A data frame containing the input data for the model.

calibratableParams

A list of calibratable parameters.

constParams

A list of constant parameters.

nw

A integer indicating the period you want to average the recharge.

warmup

A integer indicating the number of days to warm up the model.

Value

A SGD_ESTIMATION_DF class.

Details

Estimate the SGD volume based on Strack (1976) analytical solution.

The inputData should be a data.frame containing the required columns for the model (R, E0, q1, GWL, H2O_SB1, H2O_SB2).

R is a numeric vector of the daily precipitation (mm).

E0 is a numeric vector of the daily potential evapotranspiration (mm).

q1 is a numeric vector of the daily pumping rate (m3/d).

GWL is a numeric vector of the initial groundwater level (mm), Only the first value is used as the initial value for the SB1.

H2O_SB1 is a numeric vector of the initial water level in the first soil bucket (mm). Only the first value is used as the initial value for the SB1.

H2O_SB2 is a numeric vector of the inital water level in the second soil bucket (mm). Only the first value is used as the initial value for the SB2.

calibratableParams should be a list of calibratable parameters. It's recommended to use json_to_paramter_list to create this list.

nw A integer indicating the period you want to average the recharge.

warmup A integer indicating the number of days to warm up the model.