A wrapper function to call the model and return result for calibration.

A wrapper function to call the model and return result for calibration.

Usage

estimate_sgd_from_pars(
  pars,
  parnames,
  parset,
  skeleton,
  inputDf,
  yearlyPumping,
  pumpingDf,
  ...
)

Arguments

pars

A numeric vector of parameters to calibrate.

parnames

A character vector of names of the parameters.

parset

A character vector of names of the to-be-calibrated parameters.

skeleton

A named list of the parameter skeleton.

inputDf

A data.frame of the input data. See estimate_sgd for details.

yearlyPumping

A data.frame of the yearly pumping data. (2 columns: year, pumping)

pumpingDf

A data.frame of the pumping data.

...

Additional arguments passed to the model function

Value

A SGD_ESTIMATION_DF class

See also