Plot the SGD estimation results

Plot the SGD estimation results

Usage

plot(x, y, ...)

# S3 method for SGD_ESTIMATION_DF
plot(
  x,
  y,
  vars = c("wl", "SGD", "Wrechg"),
  base_date,
  type = c("pred", "obs", "comp", "input"),
  obs_x = "date",
  obs_y = "wl",
  xbreaks = 10,
  ybreaks = 5,
  maxRange = 600,
  coeff = 0.3,
  names = c("wl", "SGD", "hn", "xn", "Wrechg", "WrechgAve"),
  colors = c(obs = "red", precp = "blue", pumping = "green", est = "blue"),
  color_labels = c(obs = "Observed GWL", precp = "Precipitation", pumping = "Pumping",
    est = "Estimated GWL"),
  labels = c("WL~(m~AHD)", "SGD~(m^3/d)", "h[n]~(m~AHD)", "x[n]~(m)", "w[recharge]~(mm)",
    "w[recharge[ave]]~(mm)"),
  ...
)

Arguments

x

An object of class sgdEstimation. Created by estimate_sgd.

y

A data.frame of daily observed groundwater level data. See Details.

base_date

A valid string that can be transformed into data format using ymd.

type

A character string indicating the type of plot. Can be one of 'pred', 'obs', 'comp', 'input'.

obs_x

A character string indicating the column name of the date in the observed data.

obs_y

A character string indicating the column name of the groundwater level in the observed data.

xbreaks

An integer indicating the number of breaks on the x-axis.

ybreaks

An integer indicating the number of breaks on the y-axis.

maxRange

A numeric indicating the maximum range of the first axis (groundwater level).

coeff

A numeric indicating the shrink coefficient of the precipitation.

names

A character vector of the variable names to plot.

colors

A named character vector of the colors to use

color_labels

A named character vector of the labels for the colors.

labels

A character vector of the variable labels to plot.